Rename paths mentioning scenarios
