+++ sdk/compiler/damlc/BUILD.bazel
@@ -41,7 +41,7 @@ damlc_data = [
     "//compiler/damlc/pkg-db",
     "//compiler/damlc/stable-packages",
     "//compiler/damlc:daml-base-anchors.json",
-    "//compiler/scenario-service/server:script_service_jar",
+    "//compiler/script-service/server:script_service_jar",
 ]
 
 add_data(
@@ -121,7 +121,7 @@ package_app(
         "//compiler/damlc/daml-ide-core:dlint.yaml",
         "//compiler/damlc/pkg-db",
         "//compiler/damlc/stable-packages",
-        "//compiler/scenario-service/server:script_service_jar",
+        "//compiler/script-service/server:script_service_jar",
         "@static_asset_d3plus//:js/d3.min.js",
         "@static_asset_d3plus//:js/d3plus.min.js",
     ],
@@ -249,9 +249,9 @@ da_haskell_library(
         "//compiler/damlc/daml-opts:daml-opts-types",
         "//compiler/damlc/daml-package-config",
         "//compiler/damlc/daml-rule-types",
-        "//compiler/scenario-service/client",
-        "//compiler/scenario-service/protos:script_service_haskell_proto",
-        "//compiler/scenario-service/protos:test_results_haskell_proto",
+        "//compiler/script-service/client",
+        "//compiler/script-service/protos:script_service_haskell_proto",
+        "//compiler/script-service/protos:test_results_haskell_proto",
         "//daml-assistant:daml-lib",
         "//daml-assistant:daml-project-config",
         "//daml-assistant/daml-helper:daml-helper-lib",
