+++ sdk/compiler/damlc/tests/BUILD.bazel
@@ -163,7 +163,7 @@ da_haskell_library(
         "//compiler/damlc/daml-opts:daml-opts-types",
         "//compiler/damlc/daml-package-config",
         "//compiler/damlc/daml-rule-types",
-        "//compiler/scenario-service/client",
+        "//compiler/script-service/client",
         "//daml-assistant:daml-project-config",
         "//daml-lf/archive:daml_lf_archive_haskell_proto",
         "//libs-haskell/bazel-runfiles",
@@ -195,7 +195,7 @@ da_haskell_library(
             "//compiler/damlc:hpp-copy",
             "//compiler/damlc/pkg-db",
             "//compiler/damlc/stable-packages",
-            "//compiler/scenario-service/server:script_service_jar",
+            "//compiler/script-service/server:script_service_jar",
             "//daml-script/daml:daml-script-{version}.dar".format(version = version),
             "@jq_dev_env//:jq",
             ghc_pkg,
@@ -318,7 +318,7 @@ da_haskell_test(
     data = [
         "//compiler/damlc/pkg-db",
         "//compiler/damlc/stable-packages",
-        "//compiler/scenario-service/server:script_service_jar",
+        "//compiler/script-service/server:script_service_jar",
         "//daml-script/daml:daml-script-2.1.dar",
         ghc_pkg,
     ],
@@ -340,7 +340,7 @@ da_haskell_test(
         "//compiler/damlc/daml-ide-core",
         "//compiler/damlc/daml-ide-core:ide-testing",
         "//compiler/damlc/daml-opts:daml-opts-types",
-        "//compiler/scenario-service/client",
+        "//compiler/script-service/client",
         "//libs-haskell/da-hs-base",
         "//libs-haskell/test-utils",
         "//sdk-version/hs:sdk-version-lib",
@@ -952,7 +952,7 @@ da_haskell_test(
     data = [
         "//compiler/damlc/pkg-db",
         "//compiler/damlc/stable-packages",
-        "//compiler/scenario-service/server:script_service_jar",
+        "//compiler/script-service/server:script_service_jar",
         "//daml-script/daml:daml-script.dar",
         "//daml-script/daml:daml-script-2.dev.dar",
         ghc_pkg,
@@ -981,7 +981,7 @@ da_haskell_test(
         "//compiler/damlc/daml-opts:daml-opts-types",
         "//compiler/damlc/daml-package-config",
         "//compiler/damlc/daml-rule-types",
-        "//compiler/scenario-service/client",
+        "//compiler/script-service/client",
         "//daml-assistant:daml-project-config",
         "//libs-haskell/bazel-runfiles",
         "//libs-haskell/da-hs-base",
