+++ sdk/compiler/script-service/client/BUILD.bazel
@@ -1,0 +1,102 @@
+# Copyright (c) 2025 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+load("//bazel_tools:haskell.bzl", "da_haskell_library", "da_haskell_test")
+
+da_haskell_library(
+    name = "client",
+    srcs = glob(["src/**/*.hs"]),
+    hackage_deps = [
+        "async",
+        "base",
+        "binary",
+        "blaze-html",
+        "bytestring",
+        "conduit-extra",
+        "conduit",
+        "containers",
+        "cryptonite",
+        "deepseq",
+        "directory",
+        "extra",
+        "filepath",
+        "ghcide",
+        "grpc-haskell",
+        "grpc-haskell-core",
+        "hashable",
+        "lens",
+        "mtl",
+        "process",
+        "proto3-suite",
+        "proto3-wire",
+        "random",
+        "split",
+        "stm",
+        "text",
+        "time",
+        "transformers",
+        "uri-encode",
+        "vector",
+    ],
+    src_strip_prefix = "src",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//compiler/daml-lf-ast",
+        "//compiler/daml-lf-proto-encode",
+        "//compiler/daml-lf-proto-util",
+        "//compiler/damlc/daml-opts:daml-opts-types",
+        "//compiler/script-service/protos:script_service_haskell_proto",
+        "//daml-assistant:daml-project-config",
+        "//libs-haskell/bazel-runfiles",
+        "//libs-haskell/da-hs-base",
+    ],
+)
+
+da_haskell_test(
+    name = "tests",
+    srcs = glob(["test/**/*.hs"]),
+    hackage_deps = [
+        "async",
+        "base",
+        "binary",
+        "blaze-html",
+        "bytestring",
+        "conduit-extra",
+        "conduit",
+        "containers",
+        "cryptonite",
+        "deepseq",
+        "directory",
+        "extra",
+        "filepath",
+        "grpc-haskell",
+        "grpc-haskell-core",
+        "hashable",
+        "lens",
+        "mtl",
+        "process",
+        "proto3-suite",
+        "proto3-wire",
+        "split",
+        "stm",
+        "text",
+        "tasty",
+        "tasty-hunit",
+        "time",
+        "transformers",
+        "uri-encode",
+        "vector",
+    ],
+    main_function = "DA.Daml.LF.PrettyScriptSpec.main",
+    visibility = ["//visibility:public"],
+    deps = [
+        ":client",
+        "//compiler/daml-lf-ast",
+        "//compiler/daml-lf-proto",
+        "//compiler/damlc/daml-opts:daml-opts-types",
+        "//compiler/script-service/protos:script_service_haskell_proto",
+        "//daml-assistant:daml-project-config",
+        "//libs-haskell/bazel-runfiles",
+        "//libs-haskell/da-hs-base",
+    ],
+)
