+++ sdk/compiler/script-service/server/BUILD.bazel
@@ -1,0 +1,52 @@
+# Copyright (c) 2025 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+load("//bazel_tools:scala.bzl", "da_scala_binary", "lf_scalacopts_stricter")
+
+genrule(
+    name = "script_service_jar",
+    srcs = [":script-service-raw_distribute.jar"],
+    outs = ["script-service.jar"],
+    cmd = "cp $< $@",
+    visibility = ["//visibility:public"],
+)
+
+da_scala_binary(
+    name = "script-service-raw",
+    srcs = glob(["src/main/scala/**/*.scala"]),
+    main_class = "com.digitalasset.daml.lf.script.ScriptServiceMain",
+    resources = glob(["src/main/resources/*"]),
+    scala_deps = [
+        "@maven//:com_github_scopt_scopt",
+        "@maven//:com_typesafe_scala_logging_scala_logging",
+        "@maven//:org_apache_pekko_pekko_actor",
+        "@maven//:org_apache_pekko_pekko_stream",
+        "@maven//:org_scalaz_scalaz_core",
+        "@maven//:org_typelevel_paiges_core",
+    ],
+    scalacopts = lf_scalacopts_stricter,
+    runtime_deps = [
+        "@maven//:ch_qos_logback_logback_classic",
+    ],
+    deps = [
+        "//compiler/script-service/protos:script_service_java_proto",
+        "//daml-lf/archive:daml_lf_archive_reader",
+        "//daml-lf/data",
+        "//daml-lf/interpreter",
+        "//daml-lf/language",
+        "//daml-lf/ide-ledger",
+        "//daml-lf/stable-packages",
+        "//daml-lf/transaction",
+        "//daml-lf/validation",
+        "//daml-script/converter",
+        "//daml-script/runner:script-runner-lib",
+        "//libs-scala/contextualized-logging",
+        "//libs-scala/rs-grpc-bridge",
+        "//libs-scala/rs-grpc-pekko",
+        "@maven//:com_google_protobuf_protobuf_java",
+        "@maven//:io_grpc_grpc_api",
+        "@maven//:io_grpc_grpc_netty",
+        "@maven//:io_grpc_grpc_stub",
+        "@maven//:org_slf4j_slf4j_api",
+    ],
+)
