--- sdk/daml-lf/scenario-interpreter/BUILD.bazel
@@ -1,55 +1,0 @@
-# Copyright (c) 2025 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-load(
-    "//bazel_tools:scala.bzl",
-    "da_scala_benchmark_jmh",
-    "da_scala_library",
-    "da_scala_test",
-    "da_scala_test_suite",
-    "lf_scalacopts",
-    "lf_scalacopts_stricter",
-)
-load(
-    "//rules_daml:daml.bzl",
-    "daml_compile",
-)
-load("@scala_version//:index.bzl", "scala_major_version")
-
-da_scala_library(
-    name = "scenario-interpreter",
-    srcs = glob(["src/main/**/*.scala"]),
-    main_class = "com.daml.lf.speedy.Main",
-    scala_deps = ["@maven//:org_typelevel_paiges_core"],
-    scalacopts = lf_scalacopts_stricter,
-    tags = ["maven_coordinates=com.daml:daml-lf-scenario-interpreter:__VERSION__"],
-    visibility = [
-        "//visibility:public",
-    ],
-    deps = [
-        "//daml-lf/data",
-        "//daml-lf/engine",
-        "//daml-lf/interpreter",
-        "//daml-lf/language",
-        "//daml-lf/transaction",
-        "//libs-scala/contextualized-logging",
-        "//libs-scala/scala-utils",
-    ],
-)
-
-da_scala_test_suite(
-    name = "scenario-interpreter_tests",
-    size = "small",
-    srcs = glob(["src/test/**/*.scala"]),
-    scala_deps = ["@maven//:org_scalaz_scalaz_core"],
-    scalacopts = lf_scalacopts,
-    deps = [
-        ":scenario-interpreter",
-        "//daml-lf/data",
-        "//daml-lf/engine",
-        "//daml-lf/interpreter",
-        "//daml-lf/language",
-        "//daml-lf/transaction",
-        "//libs-scala/contextualized-logging",
-    ],
-)
